OpenSearch/pyproject.toml
[project]
name = "opensearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opensearch-py>=2.7.1",
    "python-dotenv>=1.0.1",
]
